Move component and style file to own dir
