+++ frontend/src/metabase/query_builder/components/view/ViewSidebar/ViewSidebar.styled.jsx
@@ -21,6 +21,10 @@ export const ViewSidebarAside = styled.aside`
           right: 0;
         `}
 
+  @media (prefers-reduced-motion) {
+    transition: none;
+  }
+
   ${({ isOpen, widthProp: width }) =>
     isOpen &&
     width &&
